Partial Panic!!
Just kidding.

EAGERFloodlight wasn't as extendable as I originally thought and could not actually handle multiple randomized servers. This had to do with the way prefixes were stored and chosen. Basically, the list of prefixes was stored in the Randomizer module, without any regard to which prefixes belong to which server. This was an easy fix – moving the prefix list to the server class – but it broke a lot of existing functionality; particularly the REST API.

All of this functionality should be maintained with this commit, albeit with a few changes.