The beginning of what seems to be a lot of changes. On the most basic level, we've redefined what a Connection is.
A connection used to symbolize an entire two way connection between a host and a randomized server. In order to get two way randomization, however, we now think of a connection between a source and destination (server for now), and between those two, only a single direction. This means that for every 'old' connection, there will be two 'new' connections.

This should simplify a lot of the logic – namely the flow factory.

The logic in the receive function has been pulled out into methods, simplifying the logic in the receive function.

As it stands now, this implementation is only half complete and is, therefore, broken.